libxl: Fix races in libxl event code
The libxl driver is racy in it's interactions with libxl and libvirt's
event loop.  The event loop can invoke callbacks after libxl has
deregistered the event, and possibly access freed data associated with
the event.  The race is described in more detail on the xen-devel mailing
list

  http://lists.xen.org/archives/html/xen-devel/2012-11/msg01016.html

This patch attempts to fix the race by reference counting the objects
associated with the events.  When libxl registers an event, an object
encapsulating the event is created and it's reference count is set to
1.  When the event is injected into libvirt's event loop, another
reference is taken on the object.  When libxl deregisters the event, it's
reference count is decremented.  Once the event is removed from libvirt's
event loop, the final reference is decremented and the object is disposed.
This approach ensures the object is not disposed until it is removed
from libvirt's event loop *and* libxl had explicitly deregistered the event.

The notion of an event being 'disabled' found in libvirt's event loop impl
is also added to the libxl event object via the 'registered' field.  This
ensures the driver does not make calls into libxl for a previously
deregistered event.