libxl: Fix race between destruction of objects
It is possible to destroy and cleanup a VM, resulting in freeing the
libxlDomainObjPrivate object and associated libxl ctx, before all fds and
timeouts have been deregistered and destroyed.

Fix this race by converting libxlDomainObjPrivate to a virObject and
incrementing its reference for each fd and timeout registration.  Only when
all fds and timeouts are deregistered and destroyed will the
libxlDomainObjPrivate be destroyed.